+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.9:
 
-* Fixed service restart
+* Fixed a bug in service restart
 * Added flag for the control of the automatic activation of the bundled init script
 * Refactored the build+install recipes (thanks @sdelano!)
 * Altered the configure flags detection algorithm from consulting the OHAI plugin to 
@@ -8,7 +8,9 @@
 * Various fixes to the nxensite/nxdissite scripts (thanks @sdelano!)
 * Fixed a bug where disabling the default site would not skip installing the `default` file in sites-enabled
   hence overwriting a potential `default` file installed from the administrator (thanks @sdelano!)
-* Added support for LUA Rocks
+* Added support for LUA Rocks + a `luarock` LWRP
+* Added support for more OpenResty modules
+* Small fixes all around
 
 ## 0.1.8: